% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-discretised-scale.R
\name{as.discretised_scale}
\alias{as.discretised_scale}
\title{Create discretised versions of continuous scales}
\usage{
as.discretised_scale(scale_function)
}
\arguments{
\item{scale_function}{a scale function (e.g. \code{scale_fill_divergent})}
}
\value{
A function with the same arguments as \code{scale_function} that works with discretised
values.
}
\description{
Create discretised versions of continuous scales
}
\examples{
library(ggplot2)
scale_fill_brewer_discretised <- as.discretised_scale(scale_fill_distiller)

}
\seealso{
scale_fill_discretised
}
