% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_before_using.R
\name{check_before_using}
\alias{check_before_using}
\title{Check datasets and functions in metR}
\usage{
check_before_using()
}
\value{
If all datasets and function are in metR then functions returns message:
'OK! All functions and datasets are complet in metR package'. Otherwise function is stopped with message why.
}
\description{
Check if all necessary datasets and functions are in metR
}
\examples{
check_before_using()

}
