% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-placement.R
\name{label_placement_fraction}
\alias{label_placement_fraction}
\alias{label_placement_n}
\alias{label_placement_random}
\alias{label_placement_all}
\alias{label_placement_flattest}
\title{Functions to place contour labels}
\usage{
label_placement_fraction(frac = 0.5)

label_placement_n(n = 2)

label_placement_random(seed = 42, n = 1)

label_placement_all()

label_placement_flattest(n = 1, ref_angle = 0)
}
\arguments{
\item{frac}{A numeric vector with values between 0 and 1 representing
where in the contour to put labels (i.e. \code{frac = 0.5} puts labels
at the midpoint).}

\item{n}{Number of labels to put.}

\item{seed}{Seed to use for randomly choosing where to put labels.}

\item{ref_angle}{Angle (in degrees counter-clockwise from East) to try
to approximate labels.}
}
\description{
These functions compute the position of contour labels
}
