% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\docType{data}
\name{control}
\alias{control}
\title{Methylation data in chromosome 1.}
\format{A data frame with 2 148 836 rows and 6 variables}
\usage{
data(control)
}
\description{
A dataset containing the methylation data in chromosome 1.
Data for this probe are downloaded from site: http://www.neuroepigenomics.org/methylomedb/download.html. We used 4 control samples:
Control1 AC, Control2 AC, Control3 AC and Control4 AC. After downloading we summed all results on the same position and chromosome separately
from control and disease samples. The variables are as follows:
}
\details{
\itemize{
  \item chr. name of chromosome where methylation data was collected (only chromosome 1.)
  \item poz. position of chromosome where methylation data was collected
  \item no. number of sequences
  \item meth. number of sequences where cytosine was methylated
  \item unmeth. number of sequences where cytosine was unmethylated
  \item meth.rate. methylation.rate that is ratio of meth and no variables
}
}
\keyword{datasets}
