% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coriolis.R
\name{coriolis}
\alias{coriolis}
\alias{f}
\alias{coriolis.dy}
\alias{f.dy}
\title{Effects of the Earth's rotation}
\usage{
coriolis(lat)

f(lat)

coriolis.dy(lat, a = 6371000)

f.dy(lat, a = 6371000)
}
\arguments{
\item{lat}{latitude in degrees}

\item{a}{radius of the earth}
}
\description{
Coriolis and beta parameters by latitude.
}
\details{
All functions use the correct sidereal day (24hs 56mins 4.091s) instead of
the incorrect solar day (24hs) for 0.3\% more precision and 1000\% more
pedantry.
}
