% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF_methods.R
\name{cut.eof}
\alias{cut.eof}
\title{Remove some principal components.}
\usage{
\method{cut}{eof}(x, n, ...)
}
\arguments{
\item{x}{an \code{eof} object}

\item{n}{which eofs to keep}

\item{...}{further arguments passed to or from other methods}
}
\description{
Returns an \code{eof} object with just the n principal components.
}
