% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF_methods.R
\name{denormalise}
\alias{denormalise}
\alias{denormalize}
\title{Denormalise eof matrices}
\usage{
denormalise(eof, which = c("left", "right"))

denormalize(eof, which = c("left", "right"))
}
\arguments{
\item{eof}{an \code{eof} object.}

\item{which}{which side of the eof decomposition to denormalise}
}
\description{
The matrices returned by \code{\link[=EOF]{EOF()}} are normalized. This function multiplies the left or right
matrix by the diagonal matrix to return it to proper units.
}
