% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivate.R
\name{DerivatePhysical}
\alias{DerivatePhysical}
\title{Zonal or meridional derivative}
\usage{
DerivatePhysical(variable, lon, lat, order = c(1, 2), bc = c("cyclic",
  "none"), a = 6731)
}
\arguments{
\item{variable}{numeric vector of discrete values}

\item{lon}{numeric vector of longitude (in degrees)}

\item{lat}{numeric vector of latitude (in degrees)}

\item{order}{order of the derivative}

\item{bc}{boundary conditions}
}
\value{
A numeric vector of the same length as \code{variable}.
}
\description{
Derivates a variable in the zonal or meridional direction taking into account
the size of th Earth.
}
\seealso{
\code{\link{Derivate}}
}
