% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivideTimeseries.R
\name{DivideTimeseries}
\alias{DivideTimeseries}
\title{Divides long timeseries for better reading}
\usage{
DivideTimeseries(g, x, n = 2, xlab = "x", ylab = "y")
}
\arguments{
\item{g}{ggplot object}

\item{x}{The vector that was used in g for the x axis (must be of class Date)}

\item{n}{Number of panels}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
Draws a plot.
}
\description{
Long timeseries can be compressed to the point of being unreadable when
plotted on a page. This function takes a ggplot object of a timeseries and
divides it into panels so that the time dimension gets stretched for better
readability.
}
\examples{
library(ggplot2)
series <- expand.grid(year = 1980:2000,
                     month = 1:12)
series$var = rnorm(nrow(series))
series$type = sample(letters[1:3], nrow(series), replace = TRUE)
series$date = with(series, as.Date(paste0(year, "-", month, "-01")))

g <- ggplot(series, aes(date, var, color = type)) +
    geom_line()
DivideTimeseries(g, n = 3, series$date, xlab = "Date", ylab = "Variable")
}
\seealso{
Other ggplo2 helpers: \code{\link{RepeatLon}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_x_longitude}},
  \code{\link{stat_fill_contour}}
}
