% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vector.R
\name{geom_vector}
\alias{geom_vector}
\title{Vectors}
\usage{
geom_vector(mapping = NULL, data = NULL, stat = StatVector,
  position = "identity", ..., scale = 1, scale.x = scale,
  scale.y = scale, min.mag = 0, skip = 0, skip.x = skip,
  skip.y = skip, arrow.angle = 15, arrow.length = 0.2,
  arrow.ends = "last", arrow.type = "open",
  arrow = grid::arrow(arrow.angle, unit(arrow.length, "lines"), ends =
  arrow.ends, type = arrow.type), lineend = "butt", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{scale, scale.x, scale.y}{multiplicative scales for adjusting the size of the vectors.}

\item{min.mag}{minimum magnitude of the plotted vectors}

\item{skip, skip.x, skip.y}{numeric specifying number of gridpoints not to draw in the x and y direction.}

\item{arrow.length, arrow.angle, arrow.ends, arrow.type}{parameters passed to \link[grid:arrow]{grid::arrow})}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{lineend}{Line end style (round, butt, square).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
Parametrization of \link[ggplot2:geom_segment]{ggplot2::geom_segment} by location and displacement that
also defaults to drawing arrows at the end of the segment. Useful for plotting
vector fields characterized by magnitude of the \code{x} and \code{y} components.
}
\examples{
speed <- expand.grid(lon = 1:360, lat = -90:0)
speed <- transform(speed,
                   vx = 1,
                   vy = 0.3)
library(ggplot2)
(g <- ggplot(speed, aes(lon, lat)) +
    geom_vector(aes(dx = vx, dy = vy), scale = 5,
    skip = 7, skip.x = 8))

# Every parameter (except skip) can be mapped to data.
# this can be usefull for masking vectors with a variable
# min.mag.
 ggplot(speed, aes(lon, lat)) +
    geom_vector(aes(dx = vx, dy = vy, scale.x = 5*cos(lon*pi/180)),
    skip = 7)

# Note that the segments respond to coordinate
g + coord_polar()

# Sverdrup transport
library(data.table)
b <- 10
d <- 10
grid <- as.data.table(expand.grid(x = seq(1, d, by = 0.5),
                                  y = seq(1, b, by = 0.5)))
grid[, My := -sin(pi*y/b)*pi/b]
grid[, Mx := -pi^2/b^2*cos(pi*y/b)*(d - x)]

ggplot(grid, aes(x, y)) +
    geom_vector(aes(dx = Mx, dy = My))

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_label_contour}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}}, \code{\link{stat_filter}},
  \code{\link{stat_na}}
}
\concept{ggplot2 helpers}
