% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanaka.R
\docType{data}
\name{GeomContourTanaka}
\alias{GeomContourTanaka}
\title{\code{light} and \code{dark} must be valid colours determining the light and dark shading
(defaults to "white" and "gray20", respectively). \code{sun.angle} is the angle,
in degrees counterclockwise from 12 o' clock, from which the sun is shining
(defaults to 60).}
\format{An object of class \code{GeomContourTanaka} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
GeomContourTanaka
}
\description{
\code{light} and \code{dark} must be valid colours determining the light and dark shading
(defaults to "white" and "gray20", respectively). \code{sun.angle} is the angle,
in degrees counterclockwise from 12 o' clock, from which the sun is shining
(defaults to 60).
}
\keyword{datasets}
