% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopotential.R
\docType{data}
\name{geopotential}
\alias{geopotential}
\title{Geopotential height}
\format{
A data.table with 53224 rows and 5 variables.
\describe{
\item{lon}{longitude in degrees}
\item{lat}{latitude in degrees}
\item{lev}{level in hPa}
\item{gh}{geopotential height in meters}
\item{date}{date}
}
}
\source{
\url{https://psl.noaa.gov/data/gridded/data.ncep.reanalysis.derived.pressure.html}
}
\usage{
geopotential
}
\description{
Monthly geopotential field at 700hPa south of 20°S from January 1990 to
December 2000.
}
\keyword{datasets}
