% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top.R
\name{get_top}
\alias{get_top}
\title{Get n top region}
\usage{
get_top(data, n, stats)
}
\arguments{
\item{data}{dataframe from output of find_DMR function with given method}

\item{n}{number of regions to return}

\item{stats}{data.frame with basic statisics about region. This is result of running get_stats function}
}
\value{
data.frame with n top region with their basic statistics
}
\description{
Get n top region from result of find_DMR function.
}
\examples{
data('schizophrenia')
control <- schizophrenia \%>\% filter(category == 'control') \%>\%
dplyr::select(-category)

disease <- schizophrenia \%>\% filter(category == 'disease') \%>\%
 dplyr::select(-category)

data <- preprocessing(control, disease)
data.tiles <- create_tiles_max_gap(data, gaps.length = 100)
data.tiles.small <- data.tiles \%>\% filter(tiles < 30)
stats <- get_stats(data.tiles.small)
dmr <- find_DMR(data.tiles.small, c('Wilcoxon', 'Ttest', 'KS', 'Reg.Log', 'Reg.Mixed', 'Reg.Corr.Mixed'))
get_top(find_DMR$Wilcoxon, 3, stats)
get_top(find_DMR$Reg.Log, 3, stats)

}
