% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Get Meteorological data}
\usage{
GetData(date, source = "SMN", bar = TRUE)
}
\description{
Get Meteorological data
}
\examples{
dates <- seq.Date(as.Date("2016-07-14"), as.Date("2017-07-14"), by = "1 day")
data <- GetData(dates)    # long!

ggplot(subset(data, station == "BASE BELGRANO II"),
       aes(date, (t.max + t.min)/2)) +
    geom_line()
}
