% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_vector.R
\name{guide_vector}
\alias{guide_vector}
\title{Reference arrow for magnitude scales}
\usage{
guide_vector(
  title = ggplot2::waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  keywidth = NULL,
  keyheight = NULL,
  direction = NULL,
  default.unit = "cm",
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  reverse = FALSE,
  order = 0,
  ...
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{direction}{A character string indicating the direction of the guide.
One of "horizontal" or "vertical."}

\item{override.aes}{A list specifying aesthetic parameters of legend key.
See details and examples.}

\item{nrow, ncol}{The desired number of rows and column of legends
respectively.}

\item{reverse}{logical. If \code{TRUE} the order of legends is reversed.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{...}{ignored.}
}
\description{
Draws a reference arrow. Highly experimental.
}
\seealso{
scale_vector
}
