% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdealGas.R
\name{IdealGas}
\alias{IdealGas}
\title{Ideal gas law}
\usage{
IdealGas(p, t, rho, R = 287.058)
}
\arguments{
\item{p}{pressure}

\item{t}{temperature}

\item{rho}{density}

\item{R}{gas constant for air}
}
\value{
Pressure, temperature or density, according to which variable was not passed
as an argument.
}
\description{
Calculate the density, pressure or temperature or air from the other two
variables.
}
\details{
Is important to take note of the units in which each variable is provided.
With the default value of \code{R}, pressure should be passed in Pascals, temperature
in Kelvin and density in \eqn{kg/m^3}. The result will be also in those units.
}
\examples{
IdealGas(1013*100, 20 + 273.15)
IdealGas(1013*100, rho = 1.15) - 273.15

}
\seealso{
Other meteorology functions: \code{\link{Derivate}},
  \code{\link{EOF}}, \code{\link{FitQsWave}},
  \code{\link{GeostrophicWind}}, \code{\link{WaveFlux}}
}
\concept{meteorology functions}
