% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon_trans.R
\name{lon_trans}
\alias{lon_trans}
\title{Switch between longitude conventions}
\usage{
lon_trans(to = c(180, 360))
}
\arguments{
\item{to}{convention to transform to.}
}
\description{
Transform longitude coordinates from and to [0, 360) and [-180, 180). Similar
to \link{ConverLongitude} but suitable for using in \link[ggplot2:scale_x_continuous]{ggplot2::scale_x_continuous} or,
better still, \link{scale_x_longitude}.
}
\examples{
library(ggplot2)

(g <- ggplot(geopotential[date == date[1]], aes(lon, lat)) +
    geom_contour2(aes(z = gh)))

g + scale_x_longitude(trans = lon_trans(180))

# Same as
g + scale_x_longitude(trans = "lon180")

}
