% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_longitude.R
\name{map_labels}
\alias{map_labels}
\alias{LonLabel}
\alias{LatLabel}
\title{Label longitude and latitude}
\usage{
LonLabel(lon, east = "°E", west = "°W", zero = "°")

LatLabel(lat, north = "°N", south = "°S", zero = "°")
}
\arguments{
\item{lon}{longitude in degrees}

\item{east, west, north, south, zero}{text to append for each quadrant}

\item{lat}{latitude in degrees}
}
\description{
Provide easy functions for adding suffixes to longitude and latitude for labelling
maps.
}
\details{
The default values are for Spanish.
}
\examples{

LonLabel(0:360)

}
\seealso{
Other ggplot2 helpers:
\code{\link[=MakeBreaks]{MakeBreaks()}},
\code{\link[=WrapCircular]{WrapCircular()}},
\code{\link[=geom_arrow]{geom_arrow()}},
\code{\link[=geom_contour2]{geom_contour2()}},
\code{\link[=geom_contour_fill]{geom_contour_fill()}},
\code{\link[=geom_label_contour]{geom_label_contour()}},
\code{\link[=geom_relief]{geom_relief()}},
\code{\link[=geom_streamline]{geom_streamline()}},
\code{\link[=guide_colourstrip]{guide_colourstrip()}},
\code{\link[=reverselog_trans]{reverselog_trans()}},
\code{\link{scale_divergent}},
\code{\link{scale_longitude}},
\code{\link[=stat_na]{stat_na()}},
\code{\link[=stat_subset]{stat_subset()}}
}
\concept{ggplot2 helpers}
