% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR-package.R
\docType{package}
\name{metR-package}
\alias{metR}
\alias{metR-package}
\title{metR: Tools for Easier Analysis of Meteorological Fields}
\description{
Many useful functions and extensions for dealing
    with meteorological data in the tidy data framework. Extends ggplot2
    for better plotting of scalar and vector fields and provides commonly
    used analysis methods in the atmospheric sciences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eliocamp/metR}
  \item Report bugs at \url{https://github.com/eliocamp/metR/issues}
}

}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (https://orcid.org/0000-0002-7742-9230)

}
\keyword{internal}
