% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_funs.R
\name{Percentile}
\alias{Percentile}
\title{Percentiles}
\usage{
Percentile(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A numeric vector of the same length as x with the percentile of each value
of x.
}
\description{
Computes percentiles.
}
\examples{
x <- rnorm(100)
p <- Percentile(x)

}
\seealso{
Other utilities:
\code{\link[=Anomaly]{Anomaly()}},
\code{\link[=JumpBy]{JumpBy()}},
\code{\link[=Mag]{Mag()}},
\code{\link{logic}}
}
\concept{utilities}
