% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing data}
\usage{
preprocessing(sample.1, sample.2)
}
\arguments{
\item{sample.1}{a data frame with methylation data}

\item{sample.2}{a data frame with methylation data}
}
\value{
data.frame with methylation values on common positions and chromosomes from sample.1 and sample.2
}
\description{
Preprocessing data for methylation analyses
}
\examples{
# data from package
data('schizophrenia')
control <- schizophrenia \%>\% filter(category == 'control') \%>\%
dplyr::select(-category)

disease <- schizophrenia \%>\% filter(category == 'disease') \%>\%
 dplyr::select(-category)

data <- preprocessing(control, disease)
head(data)
}
