% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitQsWave.R
\name{FitQsWave}
\alias{FitQsWave}
\alias{BuildQsWave}
\alias{BuildQsField}
\title{Quasi stationary waves}
\usage{
FitQsWave(y, k = 1)

BuildQsField(x, amplitude, phase, k)
}
\arguments{
\item{y}{numeric vector to transform}

\item{k}{numeric vector of wave numbers to keep or wave numbers to compute}

\item{x}{numeric vector of locations (in radians)}

\item{amplitude}{numeric vector of amplitudes}

\item{phase}{numeric vector of phases}
}
\value{
a named list of amplitude, phase and
wave number. \cr

\code{BuildQsField} returns a vector.
}
\description{
Quasi stationary waves
}
\details{
\code{FitQsWave} uses \code{\link{fft}} to make a fourier transform of the
data and then returns a list of parameters for each wave number kept.
The  amplitude (A), phase (\eqn{\phi}) and wave number (k) satisfy:
\deqn{y = \sum A cos((x - \phi)k)}
The phase is calculated so that it lies between 0 and \eqn{2\pi/k} so it
represents the location (in radians) of the first maximum of each wave number.  \cr

\code{BuildQsField} is, in a way, the opposite. It computes  the
the above mentioned sum but for only the selected wave numbers.
}
\seealso{
Other meteorology functions: \code{\link{Derivate}},
  \code{\link{EOF}}, \code{\link{WaveFlux}}
}
\concept{meteorology functions}
