% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_mixed.R
\name{reg.mixed}
\alias{reg.mixed}
\title{Logistic regression with random effects results for methylation data}
\usage{
reg.mixed(data)
}
\arguments{
\item{data}{There are two options:  1. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate.
This dataframe is result of function preprocessing.
2. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate, tiles and possible tiles.common columns. This dataframe is result of function create.tiles.min.gap or
create.tiles.fixed.length.}
}
\value{
vecotor with p.value and beta coef. from grouping variable from logistic regression with random effects or two-elemented vector of na values if something goes wrong
}
\description{
Get p.value and beta coefficient from grouping variable from logistic regression with random effects based on number of methylated and unmethylated citozines in two probes.
This function doesn't respect tiles or tiles.common column. This function is using in find.DMR and can be used sepparately.
}
\examples{
data('sample.1')
data('sample.2')
data <- preprocessing(sample.1, sample.2)
data.tiles <- create.tiles.max.gap(data, gaps.lenth = 100)
data.test <- data.tiles \%>\% filter(tiles == 1)
reg.mixed(data.test)
# or by some self-defined regions:
data.test.2 <- data.tiles \%>\% filter(chr == 'chr2', start > 171573000, poz < 171574000)
reg.mixed(data.test.2)
}
