% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatCircular.R
\name{RepeatCircular}
\alias{RepeatCircular}
\title{Repeat minimum longitude}
\usage{
RepeatCircular(x, circular = "lon", max = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{circular}{the name of the circular dimension}

\item{max}{the value of the longitude-like column that the duplicated
values should have. If \code{NULL}, defaults to max + resolution.}
}
\value{
A data.frame with the values corresponding to the minimum longitude
repeated next to the maximum.
}
\description{
ggplot's \code{\link[ggplot2]{coord_polar}} doesn't close contours in
cyclical data so the workaround is to copy the leftmost data to the right.
}
\examples{
library(ggplot2)
library(data.table)
data(geopotential)
g <- ggplot(geopotential[date == date[1]], aes(lon, lat)) +
    geom_contour(aes(z = gh)) +
    coord_polar() +
    ylim(c(-90, -10))

# This plot has problems in lon = 0
g

# But using RepeatCircular solves it.
g \%+\% RepeatCircular(geopotential[date == date[1]])

# The same behaviour is now implemented directly in geom_contour2
ggplot(geopotential[date == date[1]], aes(lon, lat)) +
    geom_contour2(aes(z = gh), circular = "x") +
    coord_polar() +
    ylim(c(-90, -10))

}
\seealso{
geom_contour2

Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{MakeBreaks}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour2}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_label_contour}},
  \code{\link{geom_relief}}, \code{\link{geom_streamline}},
  \code{\link{geom_vector}},
  \code{\link{guide_colourstrip}},
  \code{\link{map_labels}}, \code{\link{reverselog_trans}},
  \code{\link{scale_divergent}},
  \code{\link{scale_longitude}}, \code{\link{stat_na}},
  \code{\link{stat_subset}}
}
\concept{ggplot2 helpers}
