% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatLon.R
\name{RepeatLon}
\alias{RepeatLon}
\title{Repeat minimum longitude}
\usage{
RepeatLon(x, colname = "lon")
}
\arguments{
\item{x}{A data.table (or a data.frame which will be converted into a
data.table).}

\item{colname}{The name of the longitude-like column.}
}
\value{
A data.table with the values corresponding to the minimum longitude
repeated next to the maximum.
}
\description{
ggplot's \code{\link[ggplot2]{coord_polar}} doesn't close contours in cyclical data so
the workaround is to repeat the leftmost data.
}
\examples{
library(data.table)
library(ggplot2)
field <- setDT(expand.grid(lat = seq(-90, 0, by = 3.5),
                           lon = seq(0, 360, by = 3.5)))
field$z <- with(field, sin(lat*pi/180*4) + cos(lon*pi/180*5))
ggplot(RepeatLon(field), aes(lon, lat)) +
    geom_contour(aes(z = z)) +
    coord_polar()
}
\seealso{
Other ggplo2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_x_longitude}},
  \code{\link{stat_fill_contour}}
}
