% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBreaks.R
\name{robust_breaks}
\alias{robust_breaks}
\title{Robust contour breaks}
\usage{
robust_breaks(breaks = MakeBreaks(), quantiles = c(0.02, 0.98))
}
\arguments{
\item{breaks}{A function that takes the range of the data and binwidth as input
and returns the the breaks as output.}

\item{quantiles}{Numeric vector of length 2, sorted in ascending order,
with values between 0 and 1 representing the quantiles of \code{z} that
define the range used to compute breaks.}
}
\description{
Modifies a breaks function to compute breaks using a trimmed range
of the data to avoid extreme outliers stretching the scale.
}
