% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample2.R
\docType{data}
\name{sample.2}
\alias{sample.2}
\title{Small methylation sample}
\format{A data frame with 7 986 rows and 6 variables}
\usage{
data(sample.2)
}
\description{
A dataset containing a part of the methylation data from some patient's diseased tissue.
The variables are as follows:
}
\details{
\itemize{
  \item chr. name of chromosome where methylation data was collected
  \item poz. position of chromosome where methylation data was collected
  \item no. number of sequences
  \item meth. number of sequences where cytosine was methylated
  \item unmeth. number of sequences where cytosine was unmethylated
  \item meth.rate. methylation.rate that is ratio of meth and no variables
}
}
\keyword{datasets}
