% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\alias{AssignSeason}
\title{Assign seasons to months}
\usage{
season(x, hemisphere = c("south", "north"), lang = c("en", "es"))
}
\arguments{
\item{x}{A numeric vector of months or a vector of dates.}

\item{hemisphere}{A character with the hemisphere for which to assign the seasons.}

\item{lang}{Language to use.}
}
\value{
A factor vector of the same length as \code{x} with the season of each
month.
}
\description{
Assign seasons to months
}
\examples{
season(1, lang = "en", h = "n")
season(1, lang = "en")
season(as.Date("2017-01-01"))

}
