% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_contour_fill.R
\name{stat_contour_fill}
\alias{stat_contour_fill}
\title{Filled 2d contours of a 3d surface}
\usage{
stat_contour_fill(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
While ggplot2's \code{\link[ggplot2]{stat_contour}} can plot nice contours, it
doesn't work with the polygon geom. This stat makes some small manipulation
of the data to ensure that all contours are closed and also computes a new
aesthetic \code{int.level}, which differs from \code{level} (computed by
\link[ggplot2:stat_contour]{ggplot2::stat_contour} in that represents
the value of the \code{z} aesthetic \emph{inside} the contour instead of at the edge.
}
\section{Computed variables}{

\describe{
\item{int.level}{value of the interior contour}
}
}

\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
ggplot(surface, aes(Var1, Var2, z = value)) +
  stat_contour_fill() +
  geom_contour(color = "black", size = 0.1)

# If one uses level instead of int.level, one of the small
# contours near the crater disapears
ggplot(surface, aes(Var1, Var2, z = value)) +
  stat_contour_fill(aes(fill = ..level..))



}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}}
}
