% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_contourlabel.R
\name{stat_contourlabel}
\alias{stat_contourlabel}
\title{Label contours}
\usage{
stat_contourlabel(mapping = NULL, data = NULL, geom = "text",
  position = "identity", ..., min.size = 6, skip = 1, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{min.size}{minimum number of points for a contour to be labeled.}

\item{skip}{number of contours to skip}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
Draws labels on contours built with \link[ggplot2:stat_contour]{ggplot2::stat_contour}.
}
\details{
Is best used with a previous call to \link[ggplot2:stat_contour]{ggplot2::stat_contour} with the same
parameters.
}
\section{Aesthetics}{

\code{geom_contourlabel} understands the following aesthetics (required aesthetics are in bold):

\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{z}
\item \code{label}
\item \code{alpha}
\item \code{angle}
\item \code{colour}
\item \code{family}
\item \code{fontface}
\item \code{group}
\item \code{hjust}
\item \code{lineheight}
\item \code{size}
\item \code{vjust}
}
}

\examples{
library(ggplot2)
v <- data.table::melt(volcano)
ggplot(v, aes(Var1, Var2)) +
   geom_contour(aes(z = value)) +
   stat_contourlabel(aes(z = value), skip = 0)

# Small hack to avoid overlap between contours and labels
geom_contourlabel <- function(...) {
   list(stat_contourlabel(geom = "label", fill = "white",
                          label.r = unit(0, "lines"),
                          label.padding = unit(0.04, "lines"), color = NA, ...),
        stat_contourlabel(...))
 }

ggplot(v, aes(Var1, Var2)) +
   geom_contour(aes(z = value)) +
   geom_contourlabel(aes(z = value), skip = 1)

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_vector}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}}, \code{\link{stat_filter}},
  \code{\link{stat_na}}
}
\concept{ggplot2 helpers}
