% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fill_contour.R
\name{stat_fill_contour}
\alias{stat_fill_contour}
\title{Filled 2d contours of a 3d surface}
\usage{
stat_fill_contour(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
While ggplot2's  \code{\link{stat_contour}} can plot nice contours, it
doesn't work with the polygon geom.
}
\section{Computed variables}{

\describe{
 \item{int.level}{value of the interior contour}
 }
}

\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
ggplot(surface, aes(Var1, Var2, z = value)) +
  stat_fill_contour()
}
