% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\docType{data}
\name{surface}
\alias{surface}
\title{Surface height}
\format{
A data.table with 53224 rows and 5 variables.
\describe{
\item{lon}{longitude in degrees}
\item{lat}{latitude in degrees}
\item{height}{height in meters}
\item{x}{x coordinates of projection}
\item{y}{y coordinates of projection}
}
}
\usage{
surface
}
\description{
Surface height of central Argentina on a lambert grid.
}
\keyword{datasets}
