% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Air temperature}
\format{A data.table with 10512 rows and 3 variables:
\describe{
\item{lon}{longitude in degrees from 0 to 360}
\item{lat}{latitude in degrees}
\item{lev}{pressure level in hPa)}
\item{air}{air temperature in Kelvin}
}}
\source{
\url{https://www.esrl.noaa.gov/psd/data/gridded/data.ncep.reanalysis.derived.pressure.html}
}
\usage{
temperature
}
\description{
A global air temperature field for 2017-07-09.
}
\keyword{datasets}
