% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ks.R
\name{test_ks}
\alias{test_ks}
\title{Kolmogorov-Smirnov test for methylation data}
\usage{
test_ks(data)
}
\arguments{
\item{data}{There are two options:  1. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate.
This dataframe is result of function preprocessing.
2. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate, tiles and possible tiles.common columns. This dataframe is result of function create.tiles.min.gap or
create.tiles.fixed.length.}
}
\value{
p.value from K-S test or na values if something goes wrong
}
\description{
Get p.value from K-S test based on methylation rate in two probes.
This function doesn't respect tiles or tiles.common column. This function is using in find.DMR and can be used sepparately.
}
\examples{
data('schizophrenia')
control <- schizophrenia \%>\% filter(category == 'control') \%>\%
dplyr::select(-category)

disease <- schizophrenia \%>\% filter(category == 'disease') \%>\%
 dplyr::select(-category)

data <- preprocessing(control, disease)
data.tiles <- create_tiles_max_gap(data, gaps.length = 100)
data.test <- data.tiles \%>\% filter(tiles == 10)
test_ks(data.test)
# or by some self-defined regions:
data.test.2 <- data.tiles \%>\% filter(chr == 'chr1', poz > 80000, poz < 10000)
test_ks(data.test.2)
}
