% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t.R
\name{test_t}
\alias{test_t}
\title{t-Stundent test for methylation data}
\usage{
test_t(data)
}
\arguments{
\item{data}{There are two options:  1. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate.
This dataframe is result of function preprocessing.
2. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate, tiles and possible tiles.common columns. This dataframe is result of function create_tiles_min_gap or
create_tiles_fixed_length.}
}
\value{
p.value from t-Student test or na values if something goes wrong
}
\description{
Get p.value from t-Student with unequal variances based on methylation rate in two probes.
This function doesn't respect tiles or tiles.common column. This function is using in find_DMR and can be used sepparately.
}
\examples{
data('schizophrenia')
control <- schizophrenia \%>\% filter(category == 'control') \%>\%
dplyr::select(-category)

disease <- schizophrenia \%>\% filter(category == 'disease') \%>\%
 dplyr::select(-category)

data <- preprocessing(control, disease)
data.tiles <- create_tiles_max_gap(data, gaps.length = 100)
data.test <- data.tiles \%>\% filter(tiles == 10)
test_t(data.test)
# or by some self-defined regions:
data.test.2 <- data.tiles \%>\% filter(chr == 'chr1', poz > 80000, poz < 100000)
test_t(data.test.2)
}
