% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tajectory.R
\name{Trajectory}
\alias{Trajectory}
\title{Compute trajectories}
\usage{
Trajectory(formula, x0, y0, cyclical = FALSE, data = NULL, res = 2)
}
\arguments{
\item{formula}{a formula indicating dependent and independent variables
in the form of dx + dy ~ x + y + t.}

\item{x0, y0}{starting coordinates of the particles.}

\item{cyclical}{logical vector of boundary condition for x and y.}

\item{data}{optional data.frame containing the variables.}

\item{res}{resolution parameter (higher numbers increases the resolution)}
}
\description{
Computes trajectories of particles in a time-varying velocity field.
}
