% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{utilities}
\alias{utilities}
\alias{Anomaly}
\alias{Percentile}
\alias{Mag}
\alias{\%~\%}
\alias{\%b\%}
\alias{JumpBy}
\title{Helpful wrapers for commonly used functions.}
\usage{
Anomaly(x)

Percentile(x)

Mag(x, y)

x \%~\% target

x \%b\% limits

JumpBy(x, by, fill = c("skip", NA))
}
\description{
Helpful wrapers for commonly used functions.
}
