% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertLongitude.R
\name{ConvertLongitude}
\alias{ConvertLongitude}
\title{Converts between longitude conventions}
\usage{
ConvertLongitude(lon, from = c(360, 180))
}
\arguments{
\item{lon}{numeric vector of longitude}

\item{from}{numeric vector representing the convention of the input vector}
}
\value{
A numeric vector the same lenght of lon.
}
\description{
Converts longitude from [0, 360) to [-180, 180) and viceversa.
}
\examples{
library(ggplot2)
ggplot(aao[date == date[1]], aes(lon, lat, z = gh)) +
    geom_contour(color = "black") +
    geom_contour(aes(x = ConvertLongitude(lon, 360)))
}
