% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivate.R
\name{Derivate}
\alias{Derivate}
\title{Derivate a discrete variable using finite differences}
\usage{
Derivate(x, y, order = c(1, 2), bc = c("cyclic", "none"))
}
\arguments{
\item{x}{numeric vector of discrete values}

\item{y}{numeric vector of locations}

\item{order}{order of the derivative}

\item{bc}{boundary conditions}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Derivate a discrete variable using finite differences
}
\details{
Each element of the reuturn vector is an estimation of \eqn{dx/dy} (or
\eqn{d^2x/dy^2} if \code{order} = 2) by finite differences. The first and
last elements will be \code{NAs} unless cyclical boundary conditions are set
in \code{bc}.
}
\examples{
theta <- seq(0, 360, length.out = 20)*pi/180
x <- cos(theta)
dx_analytical <- -sin(theta)
dx_finitediff <- Derivate(x, theta)

plot(theta, dx_analytical, type = "l")
points(theta, dx_finitediff, col = "red")
}
\seealso{
\code{\link{DerivatePhysical}}

Other meteorology functions: \code{\link{DerivatePhysical}},
  \code{\link{EOF}}, \code{\link{FitQsWave}},
  \code{\link{WaveFlux}}
}
