% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorstrip.R
\name{guide_train.colorbar2}
\alias{guide_train.colorbar2}
\title{Methods}
\usage{
guide_train.colorbar2(guide, scale)
}
\arguments{
\item{guide, scale}{things you shouldn't touch}
}
\description{
Documentation for functions that CRAN made me document. DO NOT USE.
}
\examples{

print(rep("I will NOT use this functions", 100))

}
\keyword{internal}
