% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF.R
\name{EOF}
\alias{EOF}
\title{Empirical Orthogonal Function}
\usage{
EOF(z, lon, lat, date, n = 1, return = c("loadings", "field"))
}
\arguments{
\item{z}{numeric vector of variable}

\item{lon}{longitude (in degrees)}

\item{date}{date}

\item{n}{which singular value to compute}

\item{return}{should the function return loadings or the field?}

\item{lon}{longitude (in degrees)}
}
\value{
A data.table of loadings or fields.
}
\description{
Uses \code{\link[svd]{propack.svd}} to compute Empirical Orthogonal Functions
(aka Singular Value Decomposition).
}
