% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arrow.R
\name{geom_arrow}
\alias{geom_arrow}
\title{Arrows}
\usage{
geom_arrow(mapping = NULL, data = NULL, stat = StatArrow,
  position = "identity", ..., arrow.size = 0.5, arrow.angle = 14,
  lineend = "butt", skip = 0, min.mag = 0, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{arrow.size}{size of the arrow (see \link[grid:arrow]{grid::arrow})}

\item{arrow.angle}{angle of the arrow (see \link[grid:arrow]{grid::arrow})}

\item{skip}{numeric vector specifying number of gridpoints not to draw in
the x and y direction; recicled if not of length 2}

\item{min.mag}{minimum magnitude of the plotted vectors}
}
\description{
Parametrization of \link[ggplot2:geom_segment]{ggplot2::geom_segment} by location and displacement that
also defaults to drawing arrows at the end of the segment. Useful for plotting
vector fields characterized by magnitude of the \code{x} and \code{y} components.
}
\examples{
speed <- expand.grid(lon = 1:360, lat = -90:0)
speed <- transform(speed,
                   vx = 1,
                   vy = 0.3)

g <- ggplot(speed, aes(lon, lat)) +
    geom_arrow(aes(dx = vx, dy = vy), scale = 5,
               skip = c(7, 5), arrow.size = 0.2) +
    ylim(c(-90, 0))
g

g + coord_polar()

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}},
  \code{\link{stat_contour_fill}}
}
