% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Get Meteorological data}
\source{
https://ssl.smn.gob.ar/dpd/pron5d-calendario.php
}
\usage{
GetData(date, source = "SMN", bar = FALSE)
}
\arguments{
\item{date}{date vector of dates to fetch data}

\item{source}{ignored for now}

\item{bar}{logical object indicating whether to show a progress bar}
}
\value{
a dataframe with daily minimum and maximum temperature for all available stations
and each date.
}
\description{
Downloads minimum and maximum temperature station data from
Argentina's National Weather Service's
public access. Data availability is restricted to one year prior.
}
\examples{
\dontrun{
dates <- seq.Date(as.Date("2016-07-14"), as.Date("2017-07-14"), by = "1 day")
data <- GetData(dates, bar = FALSE)    # long!

library(ggplot2)
ggplot(subset(data, station == "BASE BELGRANO II"),
       aes(date, (t.max + t.min)/2)) +
    geom_line()
}

}
