% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSMNData.R
\name{GetSMNData}
\alias{GetSMNData}
\title{Get Meteorological data}
\source{
https://ssl.smn.gob.ar/dpd/pron5d-calendario.php
}
\usage{
GetSMNData(date, bar = FALSE)
}
\arguments{
\item{date}{date vector of dates to fetch data}

\item{bar}{logical object indicating whether to show a progress bar}
}
\value{
a dataframe with daily minimum and maximum temperature for all available stations
and each date.
}
\description{
Downloads minimum and maximum temperature station data from
Argentina's National Weather Service's
public access. Data availability is restricted to one year prior.
}
\examples{
\dontrun{
dates <- seq.Date(today() - 30, today(), by = "1 day")
data <- GetSMNData(dates, bar = FALSE)    # long!

library(ggplot2)
ggplot(subset(data, station == "BASE BELGRANO II"),
       aes(date, (t.max + t.min)/2)) +
    geom_line()
}

}
