% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSMNData.R
\name{GetSMNData}
\alias{GetSMNData}
\title{Get Meteorological data
This function is defunct.}
\usage{
GetSMNData(
  date,
  type = c("hourly", "daily", "radiation"),
  bar = FALSE,
  cache = TRUE,
  file.dir = tempdir()
)
}
\arguments{
\item{date}{date vector of dates to fetch data}

\item{type}{type of data to retrieve}

\item{bar}{logical object indicating whether to show a progress bar}

\item{cache}{logical indicating if the results should be saved on disk}

\item{file.dir}{optional directory where to save and/or retrieve data}
}
\value{
Nothing
}
\description{
Get Meteorological data
This function is defunct.
}
