% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeEOF.R
\name{ImputeEOF}
\alias{ImputeEOF}
\title{Impute missing values}
\usage{
ImputeEOF(data, formula, value.var, max.eof = NULL, min.eof = 1,
  tol = 1e-04, max.iter = 10000, validation = NULL, verbose = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{formula}{formula passed to \code{\link[data.table]{dcast}} to build
the matrix that will be used in the SVD decomposition (see details)}

\item{value.var}{name of the column whose values will be used}

\item{max.eof, min.eof}{maximum and minimum number of singular values used for
imputation}

\item{tol}{tolerance used for determining convergence}

\item{max.iter}{maximum iterations allowed for the algorithm}

\item{validation}{number of points to use in crossvalidation (defaults to the
maximum of 30 or 10\% of the non NA points)}

\item{verbose}{logical indicating whether to print progress}
}
\value{
A vector of inputed values with attributes \code{eof}, which is the number of
singular values used in the final inputation; and \code{rmse}, which is the Root
Mean Square Error calculated from crossvalidation.
}
\description{
Imputes missing values via Data Interpolating Empirical Orthogonal Functions
(DINEOF).
}
\details{
If \code{data} is a matrix, the \code{formula} argument is ignored and the function
returns a matrix.
}
\examples{
library(data.table)
data(aao)
aao <- copy(aao)
aao[, gh.t := Anomaly(gh), by = .(lat, lon)]
aao <- aao[date == date[1]]

# Add gaps to field
aao[, gh.gap := gh.t]
set.seed(42)
aao[sample(1:.N, .N*0.3), gh.gap := NA]

aao[, gh.impute := ImputeEOF(.SD, lon ~ lat,  value.var = "gh.gap",
                                    verbose = TRUE, max.iter = 2000)]

library(ggplot2)
ggplot(aao, aes(lon, lat)) +
    geom_contour(aes(z = gh.t), color = "black") +
    geom_contour(aes(z = gh.impute))

ggplot(aao[is.na(gh.gap)], aes(gh.t, gh.impute)) +
    geom_point()

}
\references{
Beckers, J.-M., Barth, A., and Alvera-Azcárate, A.: DINEOF reconstruction of clouded images including error maps – application to the Sea-Surface Temperature around Corsican Island, Ocean Sci., 2, 183-199, https://doi.org/10.5194/os-2-183-2006, 2006.
}
