% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{logic}
\alias{logic}
\alias{\%~\%}
\alias{\%b\%}
\title{Extended logical operatos}
\usage{
x \%~\% target

x \%b\% limits
}
\arguments{
\item{x, target, limits}{numeric vectors}
}
\value{
A logical vector of the same length of x.
}
\description{
Extended binary operators for easy subsetting.
}
\details{
\code{\%~\%} can be thought as a "similar" operator. It's a fuzzy version of
\code{\link{\%in\%}} in that returns \code{TRUE} for the element of \code{x}
which is the closest to any element of \code{target}. \cr

\code{\%b\%} can be thought as the "between" operator. It returns \code{TRUE}
for each element of \code{x} that is between the minimum and the maximum of
\code{limits}.
}
\examples{
set.seed(198)
x <- rnorm(100)
x[x \%~\% c(0.3, 0.5, 1)]

# Practical use case: vertical cross-section at
# approximately 36W between 50S and 50N.
cross.lon <- -34 + 360
library(ggplot2)
ggplot(nceptemperature[lon \%~\% cross.lon & lat \%b\% c(-50, 50)],
       aes(lat, lev)) +
    geom_contour(aes(z = air)) +
    scale_y_level() +
    scale_x_latitude()

}
\seealso{
Other utilities: \code{\link{Anomaly}},
  \code{\link{JumpBy}}, \code{\link{Mag}},
  \code{\link{Percentile}}
}
