% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{Mag}
\alias{Mag}
\title{Magnitude of a vector}
\usage{
Mag(x, y)
}
\arguments{
\item{x, y}{numeric vectors of vector components}
}
\value{
A numeric vector the same length as x and y that is \eqn{\sqrt(x^2 + y^2)}.
}
\description{
Computes the magnitude of a vector.
}
\details{
Helpful to save keystrokes and gain readability when computing wind
(or any other vector quantity) magnitude.
}
\seealso{
Other utilities: \code{\link{Anomaly}},
  \code{\link{JumpBy}}, \code{\link{Percentile}},
  \code{\link{logic}}
}
