% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBreaks.R
\name{MakeBreaks}
\alias{MakeBreaks}
\alias{AnchorBreaks}
\alias{robust_breaks}
\title{Functions for making breaks}
\usage{
MakeBreaks(binwidth = NULL, bins = 10, exclude = NULL)

AnchorBreaks(anchor = 0, binwidth = NULL, exclude = NULL, bins = 10)

robust_breaks(breaks = MakeBreaks(), quantiles = c(0.02, 0.98))
}
\arguments{
\item{binwidth}{width of breaks}

\item{bins}{number of bins, used if \code{binwidth = NULL}}

\item{exclude}{a vector of breaks to exclude}

\item{anchor}{anchor value}

\item{breaks}{A function that takes the range of the data and binwidth as input
and returns the the breaks as output.}

\item{quantiles}{Numeric vector of length 2, sorted in ascending order,
with values between 0 and 1 representing the quantiles of \code{z} that
define the range used to compute breaks.}
}
\value{
A function that takes a range as argument and a binwidth as an optional argument
and returns a sequence of equally spaced intervals covering the range.
For \code{robust_breaks}, a function to use as \code{breaks} argument in contour geoms.
It takes a data frame containing a column \code{z}, \code{binwidth} and \code{bin}
arguments and returns a numeric vector of breaks, possibly including
\code{-Inf} and \code{Inf}.
}
\description{
Functions that return functions suitable to use as the \code{breaks} argument in
ggplot2's continuous scales and in \link{geom_contour_fill}.

Modifies a breaks function to compute breaks using a trimmed range
of the data to avoid extreme outliers stretching the scale.
}
\details{
\code{MakeBreaks} is essentially an export of the default way
\link[ggplot2:stat_contour]{ggplot2::stat_contour} makes breaks.

\code{AnchorBreaks} makes breaks starting from an \code{anchor} value and covering
the range of the data according to \code{binwidth}.
}
\examples{

my_breaks <- MakeBreaks(10)
my_breaks(c(1, 100))
my_breaks(c(1, 100), 20)    # optional new binwidth argument ignored

MakeBreaks()(c(1, 100), 20)  # but is not ignored if initial binwidth is NULL
\dontshow{if (requireNamespace("reshape2")) withAutoprint(\{ # examplesIf}
# One to one mapping between contours and breaks
library(ggplot2)
binwidth <- 20
ggplot(reshape2::melt(volcano), aes(Var1, Var2, z = value)) +
    geom_contour(aes(color = after_stat(level)), binwidth = binwidth) +
    scale_color_continuous(breaks = MakeBreaks(binwidth))

#Two ways of getting the same contours. Better use the second one.
ggplot(reshape2::melt(volcano), aes(Var1, Var2, z = value)) +
    geom_contour2(aes(color = after_stat(level)), breaks = AnchorBreaks(132),
                  binwidth = binwidth) +
    geom_contour2(aes(color = after_stat(level)), breaks = AnchorBreaks(132, binwidth)) +
    scale_color_continuous(breaks = AnchorBreaks(132, binwidth))

# Sample data with extremes
surface <- reshape2::melt(volcano)
surface$value[sample(nrow(surface), 20)] <- 1000

# With regular breaks the outliers stretch the scale
# and prevent seeing the variability of the bulk of the data
ggplot(surface, aes(Var1, Var2, z = value)) +
  geom_contour_fill()

# Robust breaks removes those points from the break 
# computation and allow to see the variability
ggplot(surface, aes(Var1, Var2, z = value)) +
  geom_contour_fill(breaks = robust_breaks())
\dontshow{\}) # examplesIf}
}
\seealso{
Other ggplot2 helpers:
\code{\link[=WrapCircular]{WrapCircular()}},
\code{\link[=geom_arrow]{geom_arrow()}},
\code{\link[=geom_contour2]{geom_contour2()}},
\code{\link[=geom_contour_fill]{geom_contour_fill()}},
\code{\link[=geom_label_contour]{geom_label_contour()}},
\code{\link[=geom_relief]{geom_relief()}},
\code{\link[=geom_streamline]{geom_streamline()}},
\code{\link[=guide_colourstrip]{guide_colourstrip()}},
\code{\link{map_labels}},
\code{\link[=reverselog_trans]{reverselog_trans()}},
\code{\link{scale_divergent}},
\code{\link{scale_longitude}},
\code{\link[=stat_na]{stat_na()}},
\code{\link[=stat_subset]{stat_subset()}}
}
\concept{ggplot2 helpers}
