% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBreaks.R
\name{MakeBreaks}
\alias{MakeBreaks}
\title{Make breaks for contours}
\usage{
MakeBreaks(binwidth = NULL, bins = 10, exclude = NULL)
}
\arguments{
\item{binwidth}{width of breaks}

\item{bins}{number of bins, used if \code{binwidth = NULL}}

\item{exclude}{a vector of breaks to exclude}
}
\value{
A function that takes a range as argument and returns a sequence of equally
spaced intervals covering the range.
}
\description{
Essencially an export of the default way \link[ggplot2:stat_contour]{ggplot2::stat_contour} makes breaks,
it's intended to use as the \code{breaks} argument of \link[ggplot2:scale_color_continuous]{ggplot2::scale_color_continuous} or
\link[ggplot2:scale_fill_continuous]{ggplot2::scale_fill_continuous}. This way, there's a one to one mapping
between contours and breaks.
}
\examples{

my_breaks <- MakeBreaks(10)
my_breaks(range(1:100))

library(ggplot2)
binwidth <- 20
ggplot(reshape2::melt(volcano), aes(Var1, Var2, z = value)) +
    geom_contour(aes(color = ..level..), binwidth = binwidth) +
    scale_color_continuous(breaks = MakeBreaks(binwidth))

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_label_contour}},
  \code{\link{geom_vector}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}}, \code{\link{stat_filter}},
  \code{\link{stat_na}}
}
\concept{ggplot2 helpers}
