% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_longitude.R
\name{map_labels}
\alias{map_labels}
\alias{LonLabel}
\alias{LatLabel}
\title{Label longitude and latitude}
\usage{
LonLabel(lon, east = "째E", west = "째O")

LatLabel(lat, north = "째N", south = "째S")
}
\arguments{
\item{lon}{longitude in degrees}

\item{east, west, north, south}{text to append for each cuadrant}

\item{lat}{latitude in degrees}
}
\description{
Provide easy functions for adding suffixes to longitude and latitude for labeling
maps.
}
\details{
The default values are for Spanish.
}
\examples{

LonLabel(0:360)

# If you want it to speak in English.
LonLabel <- function(...) metR::LonLabel(..., west = "\\u00B0W")


}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{MakeBreaks}}, \code{\link{WrapCircular}},
  \code{\link{geom_arrow}}, \code{\link{geom_contour2}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_label_contour}},
  \code{\link{geom_relief}}, \code{\link{geom_streamline}},
  \code{\link{geom_vector}},
  \code{\link{guide_colourstrip}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_divergent}},
  \code{\link{scale_longitude}}, \code{\link{stat_na}},
  \code{\link{stat_subset}}
}
\concept{ggplot2 helpers}
