% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{Percentile}
\alias{Percentile}
\title{Percentiles}
\usage{
Percentile(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A numeric vector of the same length as x with the percentile of each value
of x.
}
\description{
Computes percentiles.
}
\examples{
x <- rnorm(100)
p <- Percentile(x)

# Number of extreme temperature station-days
library(data.table)
claris <- copy(claris)
claris[!is.na(max), percentile := Percentile(max), by = .(id, month(date))]
extreme.days <- claris[, .(n = sum(percentile > 0.90, na.rm = TRUE)),
                       by = .(year(date))]
library(ggplot2)
ggplot(extreme.days, aes(year, n)) +
    geom_line()
}
\seealso{
Other utilities: \code{\link{Anomaly}},
  \code{\link{JumpBy}}, \code{\link{Mag}},
  \code{\link{logic}}
}
