% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadNetCDF.R
\name{ReadNetCDF}
\alias{ReadNetCDF}
\title{Read NetCDF files.}
\usage{
ReadNetCDF(file, vars = NULL, list.vars = F)
}
\arguments{
\item{file}{File to read from.}

\item{vars}{A character vector with the name of the variables to read. If
\code{NULL}, then it read all the variables.}

\item{list.vars}{If \code{TRUE}, then it only list available variables instead of reading
their values.}
}
\value{
If \code{list.vars} is \code{FALSE} (de default) then it returns a
data.table in which each column is a variable and each rown an observation. \cr
Else, it returns a character vector with the names of the variables
(not the dimensions) included in the file.
}
\description{
Using the \code{\link{ncdf4}} package, it reads an .nc file. The advantage
over using \code{\link[ncdf4]{ncvar_get}} is that the output is a tidy data.table
with proper dimensions.
}
\examples{
\dontrun{
file <- "file.nc"
# Get a list of variables.
variables <- ReadNetCDF(file, list.vars = T)
# Read only the first one.
field <- ReadNetCDF(file, vars = variables[1])
}

}
