% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatLon.R
\name{RepeatLon}
\alias{RepeatLon}
\title{Repeat minimum longitude}
\usage{
RepeatLon(x, colname = "lon", maxlon = 360)
}
\arguments{
\item{x}{a data.frame}

\item{colname}{the name of the longitude-like column}

\item{maxlon}{the value of the longitude-like column that the duplicated
values should have.}
}
\value{
A data.frame with the values corresponding to the minimum longitude
repeated next to the maximum.
}
\description{
ggplot's \code{\link[ggplot2]{coord_polar}} doesn't close contours in
cyclical data so the workaround is to copy the leftmost data to the right.
}
\examples{
library(ggplot2)
slp <- copy(slp)
g <- ggplot(slp, aes(lon, lat)) +
    geom_contour(aes(z = slp)) +
    coord_polar() +
    ylim(c(-90, -10))

# This plot has problems in lon = 0
g

# But using RepeatLon solves it.
g \%+\% RepeatLon(slp)
}
\seealso{
Other ggplo2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}},
  \code{\link{stat_fill_contour}}
}
