% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample1.R
\docType{data}
\name{sample.1}
\alias{sample.1}
\title{Small methylation sample}
\format{A data frame with 5 132 rows and 6 variables}
\usage{
data(sample.1)
}
\description{
A dataset containing a part of the methylation data from some patient's healthy tissue.
The variables are as follows:
}
\details{
\itemize{
  \item chr. name of chromosome where methylation data was collected
  \item poz. position of chromosome where methylation data was collected
  \item no. number of reads per position
  \item meth. number of reads with methylated cytosine
  \item unmeth. number of reads with unmethylated cytosine
  \item meth.rate. methylation.rate that is ratio meth./(meth. + unmeth.)
}
}
\keyword{datasets}
