% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_divergent.R
\name{scale_color_divergent}
\alias{scale_color_divergent}
\alias{scale_fill_divergent}
\title{Divergent color scales}
\usage{
scale_color_divergent(low = scales::muted("blue"),
  high = scales::muted("red"), binwidth = NA, exclude = NULL, ...)

scale_fill_divergent(low = scales::muted("blue"),
  high = scales::muted("red"), binwidth = NA, exclude = NULL, ...)
}
\arguments{
\item{binwidth}{optional binwidth for computing breaks. For best results,
should be the same value as the related \code{stat_contour} call.}

\item{exclude}{a numeric vector of levels that should be excluded from the
contour calculation}

\item{...}{other arguments passed to \link[ggplot2:scale_color_gradient2]{ggplot2::scale_color_gradient2}}
}
\description{
Wrapper around ggplot's \code{\link[ggplot2]{scale_colour_gradient2}} with better defaults and
automatic annotated breaks. It's main use it to label the same levels as
the contours in a plot.
}
\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
ggplot(surface, aes(Var1, Var2, z = value)) +
  geom_contour(binwidth = 30, aes(color = ..level..)) +
  scale_color_divergent(binwidth = 30)

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_label_contour}},
  \code{\link{geom_vector}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_longitude}}, \code{\link{stat_filter}},
  \code{\link{stat_na}}
}
\concept{ggplot2 helpers}
