% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_divergent.R
\name{scale_color_divergent}
\alias{scale_color_divergent}
\alias{scale_fill_divergent}
\title{Divergent color scales}
\usage{
scale_color_divergent(low = muted("blue"), high = muted("red"),
  binwidth = NA, ...)

scale_fill_divergent(low = muted("blue"), high = muted("red"),
  binwidth = NA, ...)
}
\arguments{
\item{binwidth}{(optional) Binwidth for computing breaks. For best resutls,
should be the same value as the related \code{stat_contour} call.}
}
\description{
Wraper around ggplot's \code{\link[ggplot2]{scale_colour_gradient2}} with better defaults and
automatic annotated breaks. It's main use it to label the same levels as
the contours in a plot.
}
\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
gplot(surface, aes(Var1, Var2, z = value)) +
  geom_contour(binwidth = 30, aes(color = ..level..)) +
  scale_color_divergent(binwidth = 30)

}
\seealso{
Other ggplo2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{reverselog_trans}},
  \code{\link{scale_longitude}},
  \code{\link{stat_fill_contour}}
}
