% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_divergent.R
\name{scale_color_divergent}
\alias{scale_color_divergent}
\alias{scale_fill_divergent}
\title{Divergent color scales}
\usage{
scale_color_divergent(low = scales::muted("blue"),
  high = scales::muted("red"), binwidth = NA, exclude = NULL, ...)

scale_fill_divergent(low = scales::muted("blue"),
  high = scales::muted("red"), binwidth = NA, exclude = NULL, ...)
}
\arguments{
\item{low}{Colours for low and high ends of the gradient.}

\item{high}{Colours for low and high ends of the gradient.}

\item{binwidth}{optional binwidth for computing breaks. For best resutls,
should be the same value as the related \code{stat_contour} call.}

\item{exclude}{a numeric vector of levels that should be excluded from the
contour calculation}

\item{...}{Arguments passed on to \code{continuous_scale}
\describe{
  \item{aesthetics}{The names of the aesthetics that this scale works with}
  \item{scale_name}{The name of the scale}
  \item{palette}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take}
  \item{name}{The name of the scale. Used as axis or legend title. If
\code{NULL}, the default, the name of the scale is taken from the first
mapping used for that aesthetic.}
  \item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output
}}
  \item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks.
}}
  \item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
  \item{limits}{A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum.}
  \item{rescaler}{Used by diverging and n colour gradients
(i.e. \code{\link[=scale_colour_gradient2]{scale_colour_gradient2()}}, \code{\link[=scale_colour_gradientn]{scale_colour_gradientn()}}).
A function used to scale the input values to the range \eqn{[0, 1]}.}
  \item{oob}{Function that handles limits outside of the scale limits
(out of bounds). The default replaces out of bounds values with NA.}
  \item{trans}{Either the name of a transformation object, or the
object itself. Built-in transformations include "asn", "atanh",
"boxcox", "exp", "identity", "log", "log10", "log1p", "log2",
"logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

A transformation object bundles together a transform, it's inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \code{name_trans}, e.g.
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}. You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}
  \item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
  \item{super}{The super class to use for the constructed scale}
  \item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
}}
}
\description{
Wraper around ggplot's \code{\link[ggplot2]{scale_colour_gradient2}} with better defaults and
automatic annotated breaks. It's main use it to label the same levels as
the contours in a plot.
}
\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
ggplot(surface, aes(Var1, Var2, z = value)) +
  geom_contour(binwidth = 30, aes(color = ..level..)) +
  scale_color_divergent(binwidth = 30)

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_longitude}},
  \code{\link{stat_contour_fill}}
}
