% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_longitude.R
\name{scale_x_longitude}
\alias{scale_x_longitude}
\alias{scale_y_longitude}
\title{Helpful scales for maps}
\usage{
scale_x_longitude(ticks = 60, name = "", ...)

scale_y_longitude(ticks = 60, name = "", ...)
}
\arguments{
\item{ticks}{Spacing between ticks.}
}
\description{
\code{scale_*_longitude} is a wraper around \code[ggplot2]{scale_x_continuous} with
helpful defaults for ploting longitude. It can plot labels in the
+-180 convention even when data is in the 0-360 convention.
}
\seealso{
Other ggplo2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}},
  \code{\link{scale_color_divergent}},
  \code{\link{stat_fill_contour}}
}
