% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_longitude.R
\name{scale_longitude}
\alias{scale_longitude}
\alias{scale_x_longitude}
\alias{scale_latitude}
\alias{scale_y_longitude}
\alias{scale_x_latitude}
\alias{scale_y_latitude}
\alias{scale_x_level}
\alias{scale_y_level}
\title{Helpful scales for maps}
\usage{
scale_x_longitude(ticks = 60, name = "", expand = c(0, 0), ...)

scale_y_longitude(ticks = 60, name = "", expand = c(0, 0), ...)

scale_x_latitude(ticks = 30, name = "", expand = c(0, 0), ...)

scale_y_latitude(ticks = 30, name = "", expand = c(0, 0), ...)

scale_x_level(name = "", expand = c(0, 0), trans = "reverselog", ...)

scale_y_level(name = "", expand = c(0, 0), trans = "reverselog", ...)
}
\arguments{
\item{ticks}{spacing between breaks}

\item{name}{name of the scale}

\item{expand}{a numeric vector of length two giving multiplicative and
additive expansion constants.}

\item{...}{other arguments passed to \code{\link[ggplot2]{scale_x_continuous}}}

\item{trans}{either the name of a transformation object, or the object itself}
}
\description{
These functions are simple wrappers around
\code{\link[ggplot2]{scale_x_continuous}} and
\code{\link[ggplot2]{scale_y_continuous}} with
helpful defaults for plotting longitude and latitude.
}
\details{
scale_*_longitude assumes that your data is between 0 and 360 but labels it
from -180 to 180.
}
\examples{
library(ggplot2)
ggplot(aao[date == date[1]], aes(lon, lat, z = gh)) +
    geom_contour() +
    scale_x_longitude() +
    scale_y_latitude()

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{geom_vector}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{stat_filter}}, \code{\link{stat_na}}
}
\concept{ggplot2 helpers}
