% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_contour.R
\name{scale_stroke.colour_continuous}
\alias{scale_stroke.colour_continuous}
\title{Scale for stroke.colour}
\arguments{
\item{...}{parameters sent to \link[ggplot2:scale_colour_continuous]{ggplot2::scale_color_continuous}}
}
\description{
Scale for stroke.colour
}
\keyword{internal}
