% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignSeason.R
\name{season}
\alias{season}
\alias{AssignSeason}
\title{Assign seasons to months}
\usage{
season(month, hemisphere = c("south", "north"), lang = c("es", "en"))
}
\arguments{
\item{month}{A numeric vector of months or a vector of dates.}

\item{hemisphere}{A character with the hemisphere for wich to assign the seaons.}

\item{lang}{Language to use.}
}
\value{
A factor vector of the same length as \code{month} with the season of each
month.
}
\description{
Assign seasons to months
}
\examples{
season()
}
