% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmospher.R
\name{sa_pressure}
\alias{sa_pressure}
\alias{sa_height}
\alias{sa_temperature}
\alias{height_sa_trans}
\alias{pressure_sa_trans}
\alias{axis_height}
\alias{axis_pressure}
\title{Standard atmosphere}
\usage{
sa_pressure(height)

sa_height(pressure)

sa_temperature(height)

height_sa_trans()

pressure_sa_trans()

axis_height(
  name = ggplot2::waiver(),
  breaks = scales::pretty_breaks(),
  labels = ggplot2::waiver(),
  guide = ggplot2::waiver()
)

axis_pressure(
  name = ggplot2::waiver(),
  breaks = scales::pretty_breaks(),
  labels = ggplot2::waiver(),
  guide = ggplot2::waiver()
)
}
\arguments{
\item{height}{height in meter}

\item{pressure}{pressure in pascals}

\item{name, breaks, labels, guide}{arguments passed to \code{\link[ggplot2:sec_axis]{ggplot2::sec_axis()}}}
}
\description{
Utilities to use the International Standard Atmosphere. It uses the
International Standard Atmosphere up to the tropopause (11 km by definition)
and then extends up to the 500 km using the ARDC Model Atmosphere.
}
\details{
\code{sa_pressure()}, \code{sa_height()}, \code{sa_temperature()} return, respectively,
pressure (in pascals), height (in meters) and temperature (in Kelvin).

\code{height_sa_trans()} and \code{pressure_sa_trans()} are two transformation functions
to be used as the \code{trans} argument in ggplot2 scales (e.g. \verb{scale_y_continuous(trans = "height_sa"}).

\code{axis_height()}  and \code{axis_pressure()} return a secondary axis that transforms to
height or pressure respectively to be used as ggplot2 secondary axis
(e.g. \code{scale_y_continuous(sec.axis = axis_height())}).

For convenience, and unlike the "primitive" functions, both the transformation
functions and the axis functions input and output in hectopascals and kilometers.
}
\examples{
height <- seq(0, 100*1000, by = 1*1000)

# Temperature profile that defines the standard atmosphere (in degree Celsius)
plot(sa_temperature(height) - 273.15, height, type = "l")

# Pressure profile
plot(sa_pressure(height), height, type = "l", log = "x")

# Use with ggplot2
library(ggplot2)
data <- data.frame(height = height, pressure = sa_pressure(height))

ggplot(data, aes(pressure, height)) +
  geom_path() +
  scale_y_continuous(sec.axis = axis_pressure())

# Doesn't work for now
ggplot(data, aes(pressure/100, height)) +
  geom_path() +
  scale_y_continuous(trans = "pressure_sa")


}
\references{
Standard atmosphere—Glossary of Meteorology. (n.d.). Retrieved 22 February 2021, from \url{https://glossary.ametsoc.org/wiki/Standard_atmosphere}
}
