% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_na.R
\name{stat_na}
\alias{stat_na}
\title{Filter only NA values.}
\usage{
stat_na(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)
}
\description{
Usefull for indicating or masking missing data. Ths stat just
}
\examples{
library(ggplot2)
surface <- reshape2::melt(volcano)
surface <- within(surface, value[Var1 \%b\% c(20, 30) & Var2 \%b\% c(20, 30)] <- NA)
surface[sample(1:nrow(surface), 100, replace = FALSE), 3] <- NA
ggplot(surface, aes(Var1, Var2, z = value)) +
    geom_contour_fill() +
    stat_na(aes(na = value))

}
\seealso{
Other ggplot2 helpers: \code{\link{DivideTimeseries}},
  \code{\link{RepeatLon}}, \code{\link{geom_arrow}},
  \code{\link{geom_contour_fill}},
  \code{\link{reverselog_trans}},
  \code{\link{scale_color_divergent}},
  \code{\link{scale_longitude}}
}
\concept{ggplot2 helpers}
